requirements.txt
requests
googlesearch-python
flask
python-dotenv